+++ release-notes.md
@@ -1,4 +1,13 @@
-# 0.9.51-SNAPSHOT, 2026-03-30
+# 0.9.52-SNAPSHOT, 2026-04-07
+
+- improved equality checks by unpacking left and right and reapplying the operator (Janitor.java, areEquals)
+- improved the brrr module
+- orm: allow ordering to be applied to filter searches 
+- introduce a handy "shift" method for call arguments (java side)
+- the SLFLoggingRuntime can now display a custom "context", e.g. the name of a script file
+
+
+# 0.9.51, 2026-03-30
 
 - new operator: `~=` performs case-insensitive string comparisons, e.g. ("j" ~= "J") == true.
 - new maven module: janitor-modules will collect a number of "builtin" modules, which users can include in their projects as needed